# cmd/docs/swagger.yaml
definitions:
...
  github_com_sankangkin_di-rest-api_internal_models.Product:
    properties:
      brand:
        type: string
      buyPrice:
        minimum: 1
        type: integer
      categoryId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      isActive:
        type: boolean
      productName:
        minLength: 3
        type: string
      sellPricelvl1:
        minimum: 1
        type: integer
      sellPricelvl2:
        minimum: 1
        type: integer
      uom:
        minLength: 3
        type: string
      uomId:
        minimum: 3
        type: integer
      updatedAt:
        type: string
    required:
    - buyPrice
    - productName
    - sellPricelvl1
    - sellPricelvl2
    - uom
    - uomId
    type: object
...
  github_com_sankangkin_di-rest-api_internal_models.UnitConversion:
    properties:
      baseUnit:
        type: string
      baseUnitId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      deriveUnit:
        type: string
      deriveUnitId:
        type: integer
      factor:
        minimum: 1
        type: integer
      id:
        type: integer
      productId:
        type: string
      updatedAt:
        type: string
    required:
    - baseUnit
    - baseUnitId
    - deriveUnit
    - deriveUnitId
    - factor
    - productId
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  httputil.HttpError400:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  httputil.HttpError401:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: Unauthorized
        type: string
    type: object
  httputil.HttpError500:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: internal server error
        type: string
    type: object
...
  internal_domain_product.CreateProductRequstDTO:
    properties:
      brand:
        type: string
      buyPrice:
        minimum: 1
        type: integer
      categoryId:
        type: integer
      id:
        type: string
      isActive:
        type: boolean
      productName:
        minLength: 3
        type: string
      qtyOhHand:
        type: integer
      reorderlvl:
        minimum: 1
        type: integer
      sellPriceLevel1:
        minimum: 1
        type: integer
      sellPriceLevel2:
        minimum: 1
        type: integer
      uom:
        minLength: 3
        type: string
    required:
    - buyPrice
    - categoryId
    - productName
    - qtyOhHand
    - reorderlvl
    - sellPriceLevel1
    - sellPriceLevel2
    - uom
    type: object
  internal_domain_product.UpdateProductRequstDTO:
    properties:
      brand:
        description: QtyOnHand       int    `json:"qtyOhHand" validate:"required"`
        type: string
      buyPrice:
        minimum: 1
        type: integer
      categoryId:
        type: integer
      isActive:
        type: boolean
      productName:
        minLength: 3
        type: string
      reorderlvl:
        minimum: 1
        type: integer
      sellPriceLevel1:
        minimum: 1
        type: integer
      sellPriceLevel2:
        minimum: 1
        type: integer
      uom:
        minLength: 2
        type: string
    required:
    - buyPrice
    - categoryId
    - productName
    - reorderlvl
    - sellPriceLevel1
    - sellPriceLevel2
    - uom
    type: object
...
host: localhost:5555
info:
  contact:
    email: sankaungkin@gmail.com
    name: San KaungKin
  description: This is an auto-generated API docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: REST-API with(golang fiber, google wire dependency injection)
  version: "1.0"
paths:
...
  /api/product:
    post:
      consumes:
      - application/json
      description: Create a new product with name, category, prices, and status
      parameters:
      - description: Product input data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_domain_product.CreateProductRequstDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Create new product
      tags:
      - Products
  /api/products:
    get:
      consumes:
      - application/json
      description: Fetch all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch all products
      tags:
      - Products
  /api/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete individual product
      parameters:
      - description: product Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Delete individual product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Fetch individual product by Id
      parameters:
      - description: product Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch individual product by Id
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update individual product
      parameters:
      - description: product Id
        in: path
        name: id
        required: true
        type: string
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_domain_product.UpdateProductRequstDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Update individual product
      tags:
      - Products
  /api/products/conversions/:
    get:
      consumes:
      - application/json
      description: Get unit conversions
      parameters:
      - description: product Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.UnitConversion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Get unit conversions
      tags:
      - Products
  /api/products/conversions/{id}:
    get:
      consumes:
      - application/json
      description: Get unit conversions by Id
      parameters:
      - description: product Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.UnitConversion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Get unit conversions by Id
      tags:
      - Products
  /api/products/prices/:
    get:
      consumes:
      - application/json
      description: Get all product prices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Get all product prices
      tags:
      - Products
  /api/products/prices/{id}:
    get:
      consumes:
      - application/json
      description: Fetch individual product price by Id
      parameters:
      - description: product Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch individual product price by Id
      tags:
      - Products
  /api/products/stocks:
    get:
      consumes:
      - application/json
      description: Get all product stocks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Get all product stocks
      tags:
      - Products
  /api/products/stocks/{id}:
    get:
      consumes:
      - application/json
      description: Get all product stocks By Id
      parameters:
      - description: product Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Get all product stocks By Id
      tags:
      - Products
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"